@@ -13,22 +13,24 @@
         "url": "https://github.com/heremaps/oss-review-toolkit.git"
     },
     "dependencies": {
-        "antd": "^3.19.7",
-        "memoize-one": "^5.0.4",
+        "@ant-design/icons": "^4.0.2",
+        "antd": "^4.0.1",
+        "markdown-to-jsx": "^6.11.0",
+        "memoize-one": "^5.1.1",
         "prop-types": "^15.7.2",
-        "react": "^16.8.6",
-        "react-app-polyfill": "^1.0.1",
-        "react-dom": "^16.8.6",
-        "react-inspector": "^3.0.2",
-        "react-redux": "^7.1.0",
-        "recharts": "^1.6.2",
-        "redux": "^4.0.1",
+        "react": "^16.13.0",
+        "react-app-polyfill": "^1.0.6",
+        "react-dom": "^16.13.0",
+        "react-redux": "^7.2.0",
+        "react-syntax-highlighter": "^12.2.1",
+        "recharts": "2.0.0-beta.1",
+        "redux": "^4.0.5",
         "redux-logger": "^3.0.6",
-        "redux-saga": "^1.0.1",
-        "resolve": "^1.11.1",
-        "scroll-into-view-if-needed": "^2.2.20",
+        "redux-saga": "^1.1.3",
+        "resolve": "^1.15.1",
+        "scroll-into-view-if-needed": "^2.2.24",
         "spdx-correct": "^3.1.0",
-        "spdx-license-list": "^6.0.0"
+        "spdx-license-list": "^6.1.0"
     },
     "scripts": {
         "start": "node scripts/start.js",
@@ -37,56 +39,60 @@
         "lint": "eslint ./src"
     },
     "devDependencies": {
-        "@babel/core": "7.2.2",
-        "@svgr/webpack": "4.1.0",
-        "babel-core": "7.0.0-bridge.0",
-        "babel-eslint": "10.0.1",
-        "babel-jest": "24.1.0",
-        "babel-loader": "8.0.5",
-        "babel-plugin-named-asset-import": "^0.3.0",
-        "babel-preset-react-app": "^7.0.0",
-        "bfj": "6.1.1",
-        "case-sensitive-paths-webpack-plugin": "2.2.0",
-        "chalk": "2.4.2",
-        "css-loader": "2.1.0",
-        "dotenv": "6.2.0",
-        "dotenv-expand": "4.2.0",
-        "eslint": "5.13.0",
-        "eslint-config-airbnb": "17.1.0",
-        "eslint-config-react-app": "^3.0.6",
-        "eslint-loader": "2.1.2",
+        "@babel/core": "7.8.7",
+        "@svgr/webpack": "5.2.0",
+        "@typescript-eslint/eslint-plugin": "^2.22.0",
+        "@typescript-eslint/parser": "^2.22.0",
+        "babel-eslint": "10.1.0",
+        "babel-jest": "25.1.0",
+        "babel-loader": "8.0.6",
+        "babel-plugin-named-asset-import": "^0.3.6",
+        "babel-preset-react-app": "^9.1.1",
+        "case-sensitive-paths-webpack-plugin": "2.3.0",
+        "chalk": "3.0.0",
+        "css-loader": "3.4.2",
+        "dotenv": "8.2.0",
+        "dotenv-expand": "5.1.0",
+        "eslint": "6.8.0",
+        "eslint-config-airbnb": "18.0.1",
+        "eslint-config-react-app": "^5.2.0",
+        "eslint-loader": "3.0.3",
         "eslint-plugin-babel": "5.3.0",
-        "eslint-plugin-flowtype": "3.2.1",
-        "eslint-plugin-import": "2.16.0",
-        "eslint-plugin-jsx-a11y": "6.2.1",
-        "eslint-plugin-react": "7.12.4",
-        "file-loader": "3.0.1",
+        "eslint-plugin-flowtype": "4.6.0",
+        "eslint-plugin-import": "2.20.1",
+        "eslint-plugin-jsx-a11y": "6.2.3",
+        "eslint-plugin-react": "7.19.0",
+        "eslint-plugin-react-hooks": "2.5.0",
+        "file-loader": "5.1.0",
         "fork-ts-checker-webpack-plugin-alt": "0.4.14",
-        "fs-extra": "7.0.1",
+        "fs-extra": "8.1.0",
         "html-webpack-inline-source-plugin": "1.0.0-beta.2",
-        "html-webpack-plugin": "4.0.0-beta.5",
+        "html-webpack-plugin": "4.0.0-beta.11",
         "identity-obj-proxy": "3.0.0",
-        "jest": "24.1.0",
-        "jest-pnp-resolver": "1.1.0",
-        "jest-resolve": "24.1.0",
-        "mini-css-extract-plugin": "0.5.0",
-        "optimize-css-assets-webpack-plugin": "5.0.1",
-        "pnp-webpack-plugin": "^1.2.1",
-        "postcss-flexbugs-fixes": "4.1.0",
+        "jest": "25.1.0",
+        "jest-pnp-resolver": "1.2.1",
+        "jest-resolve": "25.1.0",
+        "jest-watch-typeahead": "^0.4.2",
+        "mini-css-extract-plugin": "0.9.0",
+        "optimize-css-assets-webpack-plugin": "5.0.3",
+        "pnp-webpack-plugin": "^1.6.4",
+        "postcss-flexbugs-fixes": "4.2.0",
         "postcss-loader": "3.0.0",
-        "postcss-preset-env": "6.5.0",
-        "postcss-safe-parser": "4.0.1",
-        "react-dev-utils": "^7.0.1",
-        "sass-loader": "7.1.0",
-        "style-loader": "0.23.1",
-        "terser-webpack-plugin": "1.2.2",
-        "typescript": "^3.3.3",
-        "url-loader": "1.1.2",
-        "webpack": "4.29.3",
-        "webpack-dev-server": "3.1.14",
+        "postcss-normalize": "^8.0.1",
+        "postcss-preset-env": "6.7.0",
+        "postcss-safe-parser": "4.0.2",
+        "react-dev-utils": "^10.2.0",
+        "resolve-url-loader": "^3.1.1",
+        "sass-loader": "8.0.2",
+        "style-loader": "1.1.3",
+        "terser-webpack-plugin": "2.3.5",
+        "typescript": "^3.8.3",
+        "url-loader": "3.0.0",
+        "webpack": "4.42.0",
+        "webpack-dev-server": "3.10.3",
         "webpack-event-plugin": "^1.1.1",
-        "webpack-manifest-plugin": "2.0.4",
-        "workbox-webpack-plugin": "3.6.3"
+        "webpack-manifest-plugin": "2.2.0",
+        "workbox-webpack-plugin": "4.3.1"
     },
     "babel": {
         "presets": [
