@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 HERE Europe B.V.
+ * Copyright (C) 2019-2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,13 +18,15 @@
  */
 
 import React from 'react';
-import ObjectInspector from 'react-inspector';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import {
     Modal,
     Tabs
 } from 'antd';
+import { Light as SyntaxHighlighter } from 'react-syntax-highlighter';
+import dark from 'react-syntax-highlighter/dist/esm/styles/hljs/dark';
+import markdown from 'react-syntax-highlighter/dist/esm/languages/hljs/markdown';
 import {
     getOrtResult
 } from '../reducers/selectors';
@@ -32,26 +34,57 @@ import store from '../store';
 
 const { TabPane } = Tabs;
 
+SyntaxHighlighter.registerLanguage('markdown', markdown);
+
 const AboutModal = (props) => {
     const { webAppOrtResult } = props;
-    const { data, repository: { config } } = webAppOrtResult;
-    const { excludes } = config;
-    const { paths, projects, scopes } = excludes;
-    const parameters = Array.from(data.values()).reduce(
-        (acc, item) => {
-            acc.push(...Object.entries(item));
+    const { customData, repositoryConfiguration } = webAppOrtResult;
+    const params = Object.entries(customData);
+    const renderParamsDescriptionItems = () => {
+        const items = params.reduce((acc, [param, paramValue]) => {
+            if (typeof paramValue === 'string') {
+                acc.push([param, paramValue]);
+            } else if (Array.isArray(paramValue)) {
+                acc.push([param, paramValue.join(', ')]);
+            } else {
+                Object.entries(paramValue).forEach(([paramValueKey, value]) => {
+                    acc.push([paramValueKey, value]);
+                });
+            }
 
             return acc;
-        },
-        []
-    );
+        }, []);
+
+        return items.map(([param, value]) => (
+            <tr className="ant-descriptions-row" key={`ort-params-${param}`}>
+                <th className="ant-descriptions-item-label">
+                    {param}
+                </th>
+                <td className="ant-descriptions-item-content">
+                    {
+                        value.startsWith('http')
+                            ? (
+                                <a
+                                    href={value}
+                                    rel="noopener noreferrer"
+                                    target="_blank"
+                                >
+                                    {value}
+                                </a>
+                            )
+                            : value
+                    }
+                </td>
+            </tr>
+        ));
+    };
 
     return (
         <Modal
             footer={null}
             visible
             height="90%"
-            width="80%"
+            width="90%"
             onCancel={
                 () => {
                     store.dispatch({ type: 'APP::HIDE_ABOUT_MODAL' });
@@ -60,64 +93,35 @@ const AboutModal = (props) => {
         >
             <Tabs animated={false}>
                 {
-                    (excludes.paths.length !== 0
-                    || excludes.scopes.length !== 0)
+                    (repositoryConfiguration)
                     && (
-                        <TabPane tab="Excludes" key="ort-tabs-excludes">
-                            <ObjectInspector
-                                data={{
-                                    paths,
-                                    projects,
-                                    scopes
-                                }}
-                                expandLevel={6}
-                                name=".ort.yml"
-                                showNonenumerable={false}
-                            />
+                        <TabPane tab="Excludes (.ort.yml)" key="ort-tabs-excludes">
+                            <SyntaxHighlighter
+                                language="markdown"
+                                showLineNumbers
+                                style={dark}
+                            >
+                                {repositoryConfiguration}
+                            </SyntaxHighlighter>
                         </TabPane>
                     )
                 }
                 {
-                    parameters.length > 0
+                    customData
+                    && params
+                    && params.length > 0
                     && (
-                        <TabPane tab="Parameters" key="ort-tabs-report-params">
-                            <table className="ort-params">
-                                <tbody>
-                                    {
-                                        parameters.map(([key, value]) => {
-                                            if (value.startsWith('http')) {
-                                                return (
-                                                    <tr key={`ort-params-value${key}`}>
-                                                        <th>
-                                                            {`${key} `}
-                                                        </th>
-                                                        <td>
-                                                            <a
-                                                                href={value}
-                                                                rel="noopener noreferrer"
-                                                                target="_blank"
-                                                            >
-                                                                {value}
-                                                            </a>
-                                                        </td>
-                                                    </tr>
-                                                );
-                                            }
-
-                                            return (
-                                                <tr key={`ort-params-value-${key}`}>
-                                                    <th>
-                                                        {`${key} `}
-                                                    </th>
-                                                    <td>
-                                                        {value}
-                                                    </td>
-                                                </tr>
-                                            );
-                                        })
-                                    }
-                                </tbody>
-                            </table>
+                        <TabPane tab="Parameters" key="ort-tabs-params">
+                            {/* FIXME: Replace with Ant Descriptions once uniqe key bug is fixed */}
+                            <div className="ant-descriptions ant-descriptions-small ant-descriptions-bordered">
+                                <div className="ant-descriptions-view">
+                                    <table>
+                                        <tbody>
+                                            {renderParamsDescriptionItems()}
+                                        </tbody>
+                                    </table>
+                                </div>
+                            </div>
                         </TabPane>
                     )
                 }
@@ -129,11 +133,6 @@ const AboutModal = (props) => {
                     >
                         <div className="ort-loading-logo ort-logo" />
                     </a>
-                    <p>
-                        version
-                        {' '}
-                        {webAppOrtResult.getOrtVersion()}
-                    </p>
                     <p>
                         For documentation on how to create this report please see
                         {' '}
@@ -161,7 +160,7 @@ AboutModal.propTypes = {
     webAppOrtResult: PropTypes.object.isRequired
 };
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state) => ({
     webAppOrtResult: getOrtResult(state)
 });
 
