@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017-2019 HERE Europe B.V.
+ * Copyright (C) 2017-2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,30 +21,81 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import {
-    Col, Row, Tabs
+    Col, Row, Tabs, Timeline
 } from 'antd';
+import {
+    CheckCircleOutlined, ExclamationCircleOutlined
+} from '@ant-design/icons';
 import IssuesTable from './IssuesTable';
-import SummaryViewLicenses from './SummaryViewLicenses';
-import SummaryViewTimeline from './SummaryViewTimeline';
-import PackageCollapse from './PackageCollapse';
-import ViolationsTable from './ViolationsTable';
+import LicenseChart from './LicenseChart';
+import LicenseStatsTable from './LicenseStatsTable';
+import RuleViolationsTable from './RuleViolationsTable';
 import {
     getOrtResult,
     getSummaryDeclaredLicenses,
     getSummaryDeclaredLicensesChart,
     getSummaryDeclaredLicensesFilter,
-    getSummaryDeclaredLicensesTotal,
     getSummaryDetectedLicenses,
     getSummaryDetectedLicensesChart,
     getSummaryDetectedLicensesFilter,
-    getSummaryDetectedLicensesTotal,
+    getSummaryIssuesFilter,
+    getSummaryRuleViolationsFilter,
     getSummaryViewShouldComponentUpdate
 } from '../reducers/selectors';
 import store from '../store';
 
+const { Item } = Timeline;
 const { TabPane } = Tabs;
 
 class SummaryView extends React.Component {
+    static onChangeDeclaredLicensesTable(pagination, filters, sorter, extra) {
+        store.dispatch({
+            type: 'SUMMARY::CHANGE_DECLARED_LICENSES_TABLE',
+            payload: {
+                declaredLicensesChart: extra.currentDataSource,
+                declaredLicensesFilter: {
+                    filteredInfo: filters,
+                    sortedInfo: sorter
+                }
+            }
+        });
+    }
+
+    static onChangeDetectedLicensesTable(pagination, filters, sorter, extra) {
+        store.dispatch({
+            type: 'SUMMARY::CHANGE_DETECTED_LICENSES_TABLE',
+            payload: {
+                detectedLicensesChart: extra.currentDataSource,
+                detectedLicensesFilter: {
+                    filteredInfo: filters,
+                    sortedInfo: sorter
+                }
+            }
+        });
+    }
+
+    static onChangeIssuesTable(pagination, filters) {
+        store.dispatch({
+            type: 'SUMMARY::CHANGE_ISSUES_TABLE',
+            payload: {
+                issuesFilter: {
+                    filteredInfo: filters
+                }
+            }
+        });
+    }
+
+    static onChangeRuleViolationsTable(pagination, filters) {
+        store.dispatch({
+            type: 'SUMMARY::CHANGE_RULE_VIOLATIONS_TABLE',
+            payload: {
+                ruleViolationsFilter: {
+                    filteredInfo: filters
+                }
+            }
+        });
+    }
+
     shouldComponentUpdate() {
         const { shouldComponentUpdate } = this.props;
         return shouldComponentUpdate;
@@ -52,156 +103,352 @@ class SummaryView extends React.Component {
 
     render() {
         const {
-            licenses,
+            declaredLicensesChart,
+            declaredLicensesFilter,
+            declaredLicenseStats,
+            detectedLicensesChart,
+            detectedLicensesFilter,
+            detectedLicensesStats,
+            issuesFilter,
+            ruleViolationsFilter,
             webAppOrtResult
         } = this.props;
+
+        const {
+            declaredLicenses,
+            detectedLicenses,
+            levels,
+            packages,
+            projects,
+            scopes,
+            statistics
+        } = webAppOrtResult;
+
+        const {
+            openIssues: {
+                errors: unresolvedIssues
+            },
+            openRuleViolations: {
+                errors: unresolvedRuleViolations
+            }
+        } = statistics;
+
+        const hasUnresolvedIssues = unresolvedIssues > 0;
+        const hasUnresolvedRuleViolations = unresolvedRuleViolations > 0;
 
         return (
             <div className="ort-summary">
                 <Row>
                     <Col span={22} offset={1}>
-                        <SummaryViewTimeline webAppOrtResult={webAppOrtResult} />
+                        <Timeline className="ort-summary-timeline">
+                            <Item>
+                                Found
+                                {' '}
+                                <b>
+                                    {projects.length}
+                                </b>
+                                {' '}
+                                files defining
+                                {' '}
+                                <b>
+                                    {packages.length}
+                                </b>
+                                {' '}
+                                unique dependencies within
+                                {' '}
+                                <b>
+                                    {scopes.length}
+                                </b>
+                                {' '}
+                                scopes
+                                {
+                                    scopes && scopes.length > 0
+                                    && (
+                                        <span>
+                                            {' '}
+                                            and
+                                            {' '}
+                                            <b>
+                                                {levels.length}
+                                            </b>
+                                            {' '}
+                                            dependency levels
+                                        </span>
+                                    )
+                                }
+                            </Item>
+                            <Item>
+                                {
+                                    detectedLicenses.length === 0
+                                    && (
+                                        <span>
+                                            {' '}
+                                            Detected
+                                            {' '}
+                                            <b>
+                                                {declaredLicenses.length}
+                                            </b>
+                                            {' '}
+                                            declared licenses
+                                        </span>
+                                    )
+                                }
+                                {
+                                    detectedLicenses.length !== 0
+                                    && (
+                                        <span>
+                                            Detected
+                                            {' '}
+                                            <b>
+                                                {detectedLicenses.length}
+                                            </b>
+                                            {' '}
+                                            licenses and
+                                            {' '}
+                                            <b>
+                                                {declaredLicenses.length}
+                                            </b>
+                                            {' '}
+                                            declared licenses
+                                        </span>
+                                    )
+                                }
+                            </Item>
+                            <Item
+                                dot={
+                                    (hasUnresolvedIssues || hasUnresolvedRuleViolations > 0)
+                                        ? (<ExclamationCircleOutlined style={{ fontSize: 16 }} />)
+                                        : (<CheckCircleOutlined style={{ fontSize: 16 }} />)
+                                }
+                                color={(hasUnresolvedIssues || hasUnresolvedRuleViolations) ? 'red' : 'green'}
+                            >
+                                {
+                                    hasUnresolvedIssues && !hasUnresolvedRuleViolations
+                                    && (
+                                        <span className="ort-error">
+                                            <b>
+                                                Completed scan with
+                                                {' '}
+                                                {unresolvedIssues}
+                                                {' '}
+                                                unresolved issue
+                                                {unresolvedIssues > 1 && 's'}
+                                                {' '}
+                                                in non-excluded source code or dependencies
+                                            </b>
+                                        </span>
+                                    )
+                                }
+                                {
+                                    !hasUnresolvedIssues && hasUnresolvedRuleViolations
+                                    && (
+                                        <span className="ort-error">
+                                            <b>
+                                                Completed scan with
+                                                {' '}
+                                                {unresolvedRuleViolations}
+                                                {' '}
+                                                unresolved policy violation
+                                                {unresolvedRuleViolations > 1 && 's'}
+                                                {' '}
+                                                in non-excluded source code or dependencies
+                                            </b>
+                                        </span>
+                                    )
+                                }
+                                {
+                                    hasUnresolvedIssues && hasUnresolvedRuleViolations
+                                    && (
+                                        <span className="ort-error">
+                                            <b>
+                                                Completed scan with
+                                                {' '}
+                                                {unresolvedIssues}
+                                                {' '}
+                                                unresolved issue
+                                                {unresolvedIssues > 1 && 's'}
+                                                {' '}
+                                                and
+                                                {' '}
+                                                {unresolvedRuleViolations}
+                                                {' '}
+                                                unresolved policy violation
+                                                {unresolvedRuleViolations > 1 && 's'}
+                                                {' '}
+                                                in non-excluded source code or dependencies
+                                            </b>
+                                        </span>
+                                    )
+                                }
+                                {
+                                    !hasUnresolvedIssues && !hasUnresolvedRuleViolations
+                                    && (
+                                        <span className="ort-ok">
+                                            <b>
+                                                Completed scan successfully
+                                            </b>
+                                        </span>
+                                    )
+                                }
+                            </Item>
+                        </Timeline>
                     </Col>
                 </Row>
-                <Row>
-                    <Col span={22} offset={1}>
-                        {(webAppOrtResult.hasIssues() || webAppOrtResult.hasViolations())
-                            && (
+                {
+                    (webAppOrtResult.hasIssues() || webAppOrtResult.hasRuleViolations())
+                    && (
+                        <Row>
+                            <Col span={22} offset={1}>
                                 <Tabs tabPosition="top" className="ort-summary-issues">
-                                    { webAppOrtResult.hasViolations()
+                                    {
+                                        webAppOrtResult.hasRuleViolations()
                                         && (
                                             <TabPane
                                                 tab={(
                                                     <span>
-                                                        Violations (
-                                                        {webAppOrtResult.violations.length}
-                                                        )
+                                                        Violations
                                                     </span>
                                                 )}
                                                 key="1"
                                             >
-                                                <ViolationsTable
-                                                    expandedRowRender={
-                                                        violation => (
-                                                            <PackageCollapse
-                                                                pkg={webAppOrtResult.getPackageById(violation.pkg)}
-                                                                filterScanFindings={{
-                                                                    type: ['LICENSE'],
-                                                                    value: [violation.license]
-                                                                }}
-                                                                includeIssues
-                                                                includeScanFindings
-                                                                includeViolations={false}
-                                                                showDetails={false}
-                                                                showIssues={false}
-                                                                showLicenses
-                                                                showScanFindings
-                                                                webAppOrtResult={webAppOrtResult}
-                                                            />
-                                                        )
+                                                <RuleViolationsTable
+                                                    filter={ruleViolationsFilter}
+                                                    onChange={
+                                                        SummaryView.onChangeRuleViolationsTable
                                                     }
-                                                    showPackageColumn
-                                                    violations={webAppOrtResult.violations}
+                                                    ruleViolations={webAppOrtResult.ruleViolations}
                                                 />
                                             </TabPane>
                                         )
                                     }
-                                    { webAppOrtResult.hasIssues()
+                                    {
+                                        webAppOrtResult.hasIssues()
                                         && (
                                             <TabPane
                                                 tab={(
                                                     <span>
-                                                        Issues (
-                                                        {webAppOrtResult.issues.length}
-                                                        )
+                                                        Issues
                                                     </span>
                                                 )}
                                                 key="2"
                                             >
                                                 <IssuesTable
-                                                    expandedRowRender={
-                                                        issue => (
-                                                            <PackageCollapse
-                                                                pkg={webAppOrtResult.getPackageById(issue.pkg)}
-                                                                includeIssues={false}
-                                                                includeScanFindings={false}
-                                                                includeViolations={false}
-                                                                showDetails
-                                                                showLicenses={false}
-                                                                showPaths
-                                                                webAppOrtResult={webAppOrtResult}
-                                                            />
-                                                        )
-                                                    }
+                                                    filter={issuesFilter}
                                                     issues={webAppOrtResult.issues}
-                                                    showPackageColumn
+                                                    onChange={
+                                                        SummaryView.onChangeIssuesTable
+                                                    }
                                                 />
                                             </TabPane>
                                         )
                                     }
                                 </Tabs>
-                            )
-                        }
-                    </Col>
-                </Row>
-                <Row>
-                    <Col span={22} offset={1}>
-                        <SummaryViewLicenses
-                            data={licenses}
-                            onChangeDeclaredLicensesTable={
-                                (pagination, filters, sorter, extra) => {
-                                    store.dispatch({
-                                        type: 'SUMMARY::CHANGE_DECLARED_LICENSES_TABLE',
-                                        payload: {
-                                            declaredChart: extra.currentDataSource,
-                                            declaredFilter: {
-                                                filteredInfo: filters,
-                                                sortedInfo: sorter
-                                            }
-                                        }
-                                    });
-                                }
-                            }
-                            onChangeDetectedLicensesTable={
-                                (pagination, filters, sorter, extra) => {
-                                    store.dispatch({
-                                        type: 'SUMMARY::CHANGE_DETECTED_LICENSES_TABLE',
-                                        payload: {
-                                            detectedChart: extra.currentDataSource,
-                                            detectedFilter: {
-                                                filteredInfo: filters,
-                                                sortedInfo: sorter
-                                            }
-                                        }
-                                    });
-                                }
-                            }
-                        />
-                    </Col>
-                </Row>
+                            </Col>
+                        </Row>
+                    )
+                }
+                {
+                    (webAppOrtResult.hasDeclaredLicenses()
+                    || webAppOrtResult.hasDetectedLicenses())
+                    && (
+                        <Row>
+                            <Col span={22} offset={1}>
+                                <Tabs tabPosition="top" className="ort-summary-licenses">
+                                    {
+                                        webAppOrtResult.hasDeclaredLicenses()
+                                        && (
+                                            <TabPane
+                                                tab={(
+                                                    <span>
+                                                        Declared Licenses (
+                                                        {declaredLicenses.length}
+                                                        )
+                                                    </span>
+                                                )}
+                                                key="ort-summary-declared-licenses-table"
+                                            >
+                                                <Row>
+                                                    <Col xs={24} sm={24} md={24} lg={24} xl={9}>
+                                                        <LicenseStatsTable
+                                                            emptyText="No declared licenses"
+                                                            filter={declaredLicensesFilter}
+                                                            licenses={declaredLicenses}
+                                                            licenseStats={declaredLicenseStats}
+                                                            onChange={
+                                                                SummaryView.onChangeDeclaredLicensesTable
+                                                            }
+                                                        />
+                                                    </Col>
+                                                    <Col xs={24} sm={24} md={24} lg={24} xl={15}>
+                                                        <LicenseChart licenses={declaredLicensesChart} />
+                                                    </Col>
+                                                </Row>
+                                            </TabPane>
+                                        )
+                                    }
+                                    {
+                                        webAppOrtResult.hasDetectedLicenses()
+                                        && (
+                                            <TabPane
+                                                tab={(
+                                                    <span>
+                                                        Detected Licenses (
+                                                        {detectedLicenses.length}
+                                                        )
+                                                    </span>
+                                                )}
+                                                key="ort-summary-detected-licenses-table"
+                                            >
+                                                <Row>
+                                                    <Col xs={24} sm={24} md={24} lg={24} xl={9}>
+                                                        <LicenseStatsTable
+                                                            emptyText="No detected licenses"
+                                                            filter={detectedLicensesFilter}
+                                                            licenses={detectedLicenses}
+                                                            licenseStats={detectedLicensesStats}
+                                                            onChange={SummaryView.onChangeDetectedLicensesTable}
+                                                        />
+                                                    </Col>
+                                                    <Col xs={24} sm={24} md={24} lg={24} xl={15}>
+                                                        <LicenseChart licenses={detectedLicensesChart} />
+                                                    </Col>
+                                                </Row>
+                                            </TabPane>
+                                        )
+                                    }
+                                </Tabs>
+                            </Col>
+                        </Row>
+                    )
+                }
             </div>
         );
     }
 }
 
 SummaryView.propTypes = {
-    licenses: PropTypes.object.isRequired,
+    declaredLicensesChart: PropTypes.array.isRequired,
+    declaredLicensesFilter: PropTypes.object.isRequired,
+    declaredLicenseStats: PropTypes.array.isRequired,
+    detectedLicensesChart: PropTypes.array.isRequired,
+    detectedLicensesFilter: PropTypes.object.isRequired,
+    detectedLicensesStats: PropTypes.array.isRequired,
+    issuesFilter: PropTypes.object.isRequired,
     shouldComponentUpdate: PropTypes.bool.isRequired,
+    ruleViolationsFilter: PropTypes.object.isRequired,
     webAppOrtResult: PropTypes.object.isRequired
 };
 
-const mapStateToProps = state => ({
-    licenses: {
-        declared: getSummaryDeclaredLicenses(state),
-        declaredChart: getSummaryDeclaredLicensesChart(state),
-        declaredFilter: getSummaryDeclaredLicensesFilter(state),
-        declaredTotal: getSummaryDeclaredLicensesTotal(state),
-        detected: getSummaryDetectedLicenses(state),
-        detectedChart: getSummaryDetectedLicensesChart(state),
-        detectedFilter: getSummaryDetectedLicensesFilter(state),
-        detectedTotal: getSummaryDetectedLicensesTotal(state)
-    },
+const mapStateToProps = (state) => ({
+    declaredLicensesChart: getSummaryDeclaredLicensesChart(state),
+    declaredLicensesFilter: getSummaryDeclaredLicensesFilter(state),
+    declaredLicenseStats: getSummaryDeclaredLicenses(state),
+    detectedLicensesChart: getSummaryDetectedLicensesChart(state),
+    detectedLicensesFilter: getSummaryDetectedLicensesFilter(state),
+    detectedLicensesStats: getSummaryDetectedLicenses(state),
+    issuesFilter: getSummaryIssuesFilter(state),
     shouldComponentUpdate: getSummaryViewShouldComponentUpdate(state),
+    ruleViolationsFilter: getSummaryRuleViolationsFilter(state),
     webAppOrtResult: getOrtResult(state)
 });
 
