@@ -1,12 +1,34 @@
+/*
+ * Copyright (C) 2017-2020 HERE Europe B.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * License-Filename: LICENSE
+ */
+
 import React from 'react';
 import { connect } from 'react-redux';
 import {
     Button,
-    Icon,
+    Collapse,
     Table,
-    // Tag,
     Tooltip
 } from 'antd';
+import {
+    FileAddOutlined,
+    FileExcelOutlined
+} from '@ant-design/icons';
 import PropTypes from 'prop-types';
 import {
     getOrtResult,
@@ -14,7 +36,14 @@ import {
     getTableViewShouldComponentUpdate
 } from '../reducers/selectors';
 import store from '../store';
-import PackageCollapse from './PackageCollapse';
+import PackageDetails from './PackageDetails';
+import PackageFindingsTable from './PackageFindingsTable';
+import PackageLicenses from './PackageLicenses';
+import PackagePaths from './PackagePaths';
+import PathExcludesTable from './PathExcludesTable';
+import ScopeExcludesTable from './ScopeExcludesTable';
+
+const { Panel } = Collapse;
 
 class TableView extends React.Component {
     shouldComponentUpdate() {
@@ -33,192 +62,196 @@ class TableView extends React.Component {
             webAppOrtResult
         } = this.props;
 
+        const sortColumnFilters = (a, b) => {
+            if (a.text < b.text) {
+                return -1;
+            }
+            if (a.text > b.text) {
+                return 1;
+            }
+            return 0;
+        };
+
         // Specifies table columns as per
         // https://ant.design/components/table/
-        const columns = [
-            {
+        const columns = [];
+
+        if (webAppOrtResult.hasExcludes()) {
+            columns.push({
                 align: 'right',
                 filters: (() => [
-                    { text: 'Issues', value: 'issues' },
-                    { text: 'Violations', value: 'violations' }
+                    {
+                        text: (
+                            <span>
+                                <FileExcelOutlined className="ort-excluded" />
+                                {' '}
+                                Excluded
+                            </span>
+                        ),
+                        value: 'excluded'
+                    },
+                    {
+                        text: (
+                            <span>
+                                <FileAddOutlined />
+                                {' '}
+                                Included
+                            </span>
+                        ),
+                        value: 'included'
+                    }
                 ])(),
                 filterMultiple: true,
-                key: 'issues',
-                onFilter: (value, pkg) => {
-                    if (value === 'issues') {
-                        return pkg.hasIssues(webAppOrtResult);
+                filteredValue: filteredInfo.excludes || null,
+                key: 'excludes',
+                onFilter: (value, webAppPackage) => {
+                    if (value === 'excluded') {
+                        return webAppPackage.isExcluded;
                     }
 
-                    if (value === 'violations') {
-                        return pkg.hasViolations(webAppOrtResult);
+                    if (value === 'included') {
+                        return !webAppPackage.isExcluded;
                     }
 
                     return false;
                 },
-                render: (pkg) => {
-                    if (pkg.hasIssues(webAppOrtResult) || pkg.hasViolations(webAppOrtResult)) {
-                        return (
-                            <Icon
-                                type="exclamation-circle"
-                                className="ort-error"
-                            />
-                        );
-                    }
-
-                    return (
-                        <Icon
-                            type="check-circle"
-                            className="ort-success"
-                        />
-                    );
-                },
-                width: '0.8em'
-            },
-            {
-                align: 'right',
-                dataIndex: 'projects',
-                filters: (() => {
-                    const projects = webAppOrtResult.getProjects();
-                    return projects.map((project, index) => ({ text: project.definitionFilePath, value: index }));
-                })(),
-                filteredValue: filteredInfo.projects || null,
-                onFilter: (value, record) => record.projectIndexes.includes(parseInt(value, 10)),
-                render: (text, record) => {
-                    const prj = webAppOrtResult
-                        .getProjectByIndex(record.projectIndexes[0]);
-                    if (prj && prj.definitionFilePath) {
-                        return (
-                            <span className="ort-project-id">
-                                <Tooltip
-                                    placement="right"
-                                    title={`Defined in ${prj.definitionFilePath}`}
-                                >
-                                    <Icon type="file-text" />
-                                </Tooltip>
-                            </span>
-                        );
-                    }
-
-                    return (
-                        <span className="ort-project-id">
-                            <Icon type="file-text" />
+                render: (webAppPackage) => (
+                    webAppPackage.isExcluded ? (
+                        <span className="ort-excludes">
+                            <Tooltip
+                                placement="right"
+                                title={Array.from(webAppPackage.pathExcludeReasons).join(', ')}
+                            >
+                                <FileExcelOutlined className="ort-excluded" />
+                            </Tooltip>
                         </span>
-                    );
-                },
-                width: '0.8em'
-            },
-            {
-                align: 'left',
-                dataIndex: 'id',
-                onFilter: (value, record) => record.id.includes(value),
-                sorter: (a, b) => {
-                    const idA = a.id.toUpperCase();
-                    const idB = b.id.toUpperCase();
-                    if (idA < idB) {
-                        return -1;
-                    }
-                    if (idA > idB) {
-                        return 1;
-                    }
+                    ) : (
+                        <FileAddOutlined />
+                    )
+                ),
+                width: '2em'
+            });
+        }
 
-                    return 0;
-                },
-                sortOrder: sortedInfo.columnKey === 'id' && sortedInfo.order,
-                title: 'Package',
-                render: text => (
-                    <span
-                        className="ort-package-id ort-word-break-wrap"
-                    >
-                        {text}
+        columns.push({
+            align: 'left',
+            dataIndex: 'id',
+            ellipsis: true,
+            filters: (() => {
+                const { projects } = webAppOrtResult;
+                return projects
+                    .map(
+                        (project) => (
+                            {
+                                text: project.definitionFilePath ? project.definitionFilePath : project.id,
+                                value: project._id
+                            }
+                        )
+                    )
+                    .sort(sortColumnFilters);
+            })(),
+            filterMultiple: true,
+            filteredValue: filteredInfo.id || null,
+            onFilter: (value, webAppPackage) => webAppPackage.projectIndexes.has(value),
+            sorter: (a, b) => {
+                const idA = a.id.toUpperCase();
+                const idB = b.id.toUpperCase();
+                if (idA < idB) {
+                    return -1;
+                }
+                if (idA > idB) {
+                    return 1;
+                }
+
+                return 0;
+            },
+            sortOrder: sortedInfo.columnKey === 'id' && sortedInfo.order,
+            title: 'Package'
+        });
+
+        if (webAppOrtResult.hasScopes()) {
+            columns.push({
+                align: 'left',
+                dataIndex: 'scopeIndexes',
+                filters: (
+                    () => webAppOrtResult.scopes
+                        .map(
+                            (scope) => ({ text: scope.name, value: scope.id })
+                        )
+                        .sort(sortColumnFilters)
+                )(),
+                filteredValue: filteredInfo.scopeIndexes || null,
+                onFilter: (value, webAppPackage) => webAppPackage.hasScopeIndex(value),
+                title: 'Scopes',
+                render: (scopeIndexes, webAppPackage) => (
+                    <span>
+                        {Array.from(webAppPackage.scopeNames).join(',')}
                     </span>
                 )
-            },
-            {
-                align: 'left',
-                dataIndex: 'scopes',
-                filters: (() => webAppOrtResult.scopes.map(scope => ({ text: scope, value: scope })))(),
-                filteredValue: filteredInfo.scopes || null,
-                onFilter: (scope, component) => component.scopes.includes(scope),
-                title: 'Scopes',
-                render: (text, row) => (
-                    <ul className="ort-table-list">
-                        {row.scopes.map(scope => (
-                            <li key={`scope-${scope}`}>
-                                {scope}
-                            </li>
-                        ))}
-                    </ul>
-                )
-            },
-            {
+            });
+        }
+
+        if (webAppOrtResult.hasLevels()) {
+            columns.push({
                 align: 'center',
                 dataIndex: 'levels',
-                filters: (() => webAppOrtResult.levels.map(level => ({ text: level, value: level })))(),
+                filters: (() => webAppOrtResult.levels.map((level) => ({ text: level, value: level })))(),
                 filteredValue: filteredInfo.levels || null,
                 filterMultiple: true,
-                onFilter: (level, component) => component.levels.includes(parseInt(level, 10)),
-                render: (text, row) => (
-                    <ul className="ort-table-list">
-                        {row.levels.sort().map(level => (
-                            <li key={`level-${level}`}>
-                                {level}
-                            </li>
-                        ))}
-                    </ul>
-                ),
+                onFilter: (value, webAppPackage) => webAppPackage.hasLevel(value),
+                textWrap: 'word-break',
                 title: 'Levels',
+                render: (levels) => (
+                    <span>
+                        {Array.from(levels).join(', ')}
+                    </span>
+                ),
                 width: 80
-            },
-            {
+            });
+        }
+
+        if (webAppOrtResult.hasDeclaredLicenses()) {
+            columns.push({
                 align: 'left',
                 dataIndex: 'declaredLicenses',
-                filters: (() => webAppOrtResult.declaredLicenses.map(license => ({ text: license, value: license })))(),
+                filters: (
+                    () => webAppOrtResult.declaredLicenses.map((license) => ({ text: license, value: license }))
+                )(),
                 filteredValue: filteredInfo.declaredLicenses || null,
                 filterMultiple: true,
                 key: 'declaredLicenses',
-                onFilter: (value, record) => record.declaredLicenses.includes(value),
+                onFilter: (value, webAppPackage) => webAppPackage.declaredLicenses.has(value),
+                textWrap: 'word-break',
                 title: 'Declared Licenses',
-                render: (text, row) => (
-                    <ul className="ort-table-list">
-                        {row.declaredLicenses.map((license, index) => (
-                            <span
-                                className="ort-word-break-wrap"
-                                key={`ort-package-license-${license}`}
-                            >
-                                {license}
-                                {index !== (row.declaredLicenses.length - 1) && ', '}
-                            </span>
-                        ))}
-                    </ul>
+                render: (declaredLicenses) => (
+                    <span>
+                        {Array.from(declaredLicenses).join(', ')}
+                    </span>
                 ),
-                width: 160
-            },
-            {
+                width: '18%'
+            });
+        }
+
+        if (webAppOrtResult.hasDetectedLicenses()) {
+            columns.push({
                 align: 'left',
                 dataIndex: 'detectedLicenses',
-                filters: (() => webAppOrtResult.detectedLicenses.map(license => ({ text: license, value: license })))(),
+                filters: (
+                    () => webAppOrtResult.detectedLicenses.map((license) => ({ text: license, value: license }))
+                )(),
                 filteredValue: filteredInfo.detectedLicenses || null,
                 filterMultiple: true,
-                onFilter: (license, component) => component.detectedLicenses.includes(license),
+                onFilter: (license, webAppPackage) => webAppPackage.detectedLicenses.has(license),
+                textWrap: 'word-break',
                 title: 'Detected Licenses',
-                render: (text, row) => (
-                    <ul className="ort-table-list">
-                        {
-                            row.detectedLicenses.map((license, index) => (
-                                <span
-                                    className="ort-word-break-wrap"
-                                    key={`ort-package-license-${license}`}
-                                >
-                                    {license}
-                                    {index !== (row.detectedLicenses.length - 1) && ', '}
-                                </span>
-                            ))
-                        }
-                    </ul>
+                render: (detectedLicenses) => (
+                    <span>
+                        {Array.from(detectedLicenses).join(', ')}
+                    </span>
                 ),
-                width: 160
-            }
-        ];
+                width: '18%'
+            });
+        }
 
         return (
             <div>
@@ -235,14 +268,65 @@ class TableView extends React.Component {
                 <Table
                     columns={columns}
                     expandedRowRender={
-                        pkg => (
-                            <PackageCollapse
-                                pkg={pkg}
-                                webAppOrtResult={webAppOrtResult}
-                            />
+                        (webAppPackage) => (
+                            <Collapse
+                                className="ort-package-collapse"
+                                bordered={false}
+                                defaultActiveKey={[0, 1]}
+                            >
+                                <Panel header="Details" key="0">
+                                    <PackageDetails webAppPackage={webAppPackage} />
+                                </Panel>
+                                {
+                                    webAppPackage.hasLicenses()
+                                    && (
+                                        <Panel header="Licenses" key="1">
+                                            <PackageLicenses webAppPackage={webAppPackage} />
+                                        </Panel>
+                                    )
+                                }
+                                {
+                                    webAppPackage.hasPaths()
+                                    && (
+                                        <Panel header="Paths" key="2">
+                                            <PackagePaths paths={webAppPackage.paths} />
+                                        </Panel>
+                                    )
+                                }
+                                {
+                                    webAppPackage.hasFindings()
+                                    && (
+                                        <Panel header="Scan Results" key="3">
+                                            <PackageFindingsTable
+                                                webAppPackage={webAppPackage}
+                                            />
+                                        </Panel>
+                                    )
+                                }
+                                {
+                                    webAppPackage.hasPathExcludes()
+                                    && (
+                                        <Panel header="Path Excludes" key="4">
+                                            <PathExcludesTable
+                                                excludes={webAppPackage.pathExcludes}
+                                            />
+                                        </Panel>
+                                    )
+                                }
+                                {
+                                    webAppPackage.hasScopeExcludes()
+                                    && (
+                                        <Panel header="Scope Excludes" key="5">
+                                            <ScopeExcludesTable
+                                                excludes={webAppPackage.scopeExcludes}
+                                            />
+                                        </Panel>
+                                    )
+                                }
+                            </Collapse>
                         )
                     }
-                    dataSource={webAppOrtResult.packagesTreeFlatArray}
+                    dataSource={webAppOrtResult.packages}
                     expandRowByClick
                     indentSize={0}
                     locale={{
@@ -284,7 +368,7 @@ TableView.propTypes = {
     webAppOrtResult: PropTypes.object.isRequired
 };
 
-const mapStateToProps = state => ({
+const mapStateToProps = (state) => ({
     shouldComponentUpdate: getTableViewShouldComponentUpdate(state),
     tableView: getTableView(state),
     webAppOrtResult: getOrtResult(state)
