@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 HERE Europe B.V.
+ * Copyright (C) 2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,26 +17,30 @@
  * License-Filename: LICENSE
  */
 
-class ScannerDetails {
+class WebAppCopyright {
+    #_id;
+
+    #statement;
+
     constructor(obj) {
-        this.name = '';
-        this.version = '';
-        this.configuration = '';
-
-        if (obj instanceof Object) {
-            if (obj.name) {
-                this.name = obj.name;
+        if (obj) {
+            if (Number.isInteger(obj._id)) {
+                this.#_id = obj._id;
             }
 
-            if (obj.version) {
-                this.version = obj.version;
-            }
-
-            if (obj.configuration) {
-                this.configuration = obj.configuration;
+            if (obj.statement) {
+                this.#statement = obj.statement;
             }
         }
     }
+
+    get _id() {
+        return this.#_id;
+    }
+
+    get statement() {
+        return this.#statement;
+    }
 }
 
-export default ScannerDetails;
+export default WebAppCopyright;
