@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 HERE Europe B.V.
+ * Copyright (C) 2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,26 +17,35 @@
  * License-Filename: LICENSE
  */
 
-class ScannerDetails {
+import { UNIQUE_COLORS } from '../data/colors';
+
+class WebAppLicense {
+    #_id;
+
+    #id;
+
     constructor(obj) {
-        this.name = '';
-        this.version = '';
-        this.configuration = '';
-
-        if (obj instanceof Object) {
-            if (obj.name) {
-                this.name = obj.name;
+        if (obj) {
+            if (Number.isInteger(obj._id)) {
+                this.#_id = obj._id;
             }
 
-            if (obj.version) {
-                this.version = obj.version;
+            if (obj.id) {
+                this.#id = obj.id;
             }
 
-            if (obj.configuration) {
-                this.configuration = obj.configuration;
-            }
+            const { data: colors } = UNIQUE_COLORS;
+            this.color = colors[this.#_id % colors.length];
         }
+    }
+
+    get _id() {
+        return this.#_id;
+    }
+
+    get id() {
+        return this.#id;
     }
 }
 
-export default ScannerDetails;
+export default WebAppLicense;
