@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 HERE Europe B.V.
+ * Copyright (C) 2019-2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,38 +17,201 @@
  * License-Filename: LICENSE
  */
 
-/* eslint constructor-super: 0 */
-
-import OrtIssue from './OrtIssue';
 import { randomStringGenerator } from '../utils';
 
-class WebAppOrtIssue extends OrtIssue {
-    #key
+class WebAppOrtIssue {
+    #_id;
 
-    #pkg;
+    #message;
 
-    constructor(obj) {
+    #package;
+
+    #packageIndex;
+
+    #path;
+
+    #scanResult;
+
+    #scanResultIndex;
+
+    #severity;
+
+    #source;
+
+    #timestamp;
+
+    #type;
+
+    #resolutionIndexes = new Set();
+
+    #resolutionReasons;
+
+    #resolutions;
+
+    #webAppOrtResult;
+
+    constructor(obj, webAppOrtResult) {
         if (obj) {
-            super(obj);
-
-            if (obj.pkg) {
-                this.#pkg = obj.pkg;
+            if (Number.isInteger(obj._id)) {
+                this.#_id = obj._id;
             }
 
-            this.#key = randomStringGenerator(20);
+            if (obj.message) {
+                this.#message = obj.message;
+            }
+
+            if (obj.path) {
+                this.#path = obj.path;
+            }
+
+            if (Number.isInteger(obj.pkg)) {
+                this.#packageIndex = obj.pkg;
+            }
+
+            if (Number.isInteger(obj.scan_result) || Number.isInteger(obj.scanResult)) {
+                this.#scanResultIndex = obj.scan_result || obj.scanResult;
+            }
+
+            if (obj.severity) {
+                this.#severity = obj.severity;
+            }
+
+            if (obj.source) {
+                this.#source = obj.source;
+            }
+
+            if (obj.timestamp) {
+                this.#timestamp = obj.timestamp;
+            }
+
+            if (obj.type) {
+                this.#type = obj.type;
+            }
+
+            if (obj.resolutions) {
+                this.#resolutionIndexes = new Set(obj.resolutions);
+            }
+
+            if (webAppOrtResult) {
+                this.#webAppOrtResult = webAppOrtResult;
+
+                const webAppPackage = webAppOrtResult.getPackageByIndex(this.#packageIndex);
+                if (webAppPackage) {
+                    this.#package = webAppPackage;
+                }
+
+                const webAppScanResult = webAppOrtResult.getScanResultByIndex(this.#packageIndex);
+                if (webAppScanResult) {
+                    this.#scanResult = webAppScanResult;
+                }
+            }
+
+            this.key = randomStringGenerator(20);
         }
     }
 
-    get key() {
-        return this.#key;
+    get _id() {
+        return this.#_id;
     }
 
-    get pkg() {
-        return this.#pkg;
+    get isResolved() {
+        if (this.#resolutionIndexes && this.#resolutionIndexes.size > 0) {
+            return true;
+        }
+
+        return false;
     }
 
-    set pkg(pkg) {
-        this.#pkg = pkg;
+    get message() {
+        return this.#message;
+    }
+
+    get package() {
+        return this.#package;
+    }
+
+    get packageIndex() {
+        return this.#packageIndex;
+    }
+
+    get packageName() {
+        return this.#package ? this.#package.id : '';
+    }
+
+    get path() {
+        return this.#path;
+    }
+
+    get scanResult() {
+        return this.#scanResult;
+    }
+
+    get severity() {
+        return this.#severity;
+    }
+
+    get severityIndex() {
+        if (this.isResolved) {
+            return 3;
+        }
+
+        if (this.#severity === 'ERROR') {
+            return 0;
+        }
+
+        if (this.#severity === 'WARNING') {
+            return 1;
+        }
+
+        if (this.#severity === 'HINT') {
+            return 2;
+        }
+
+        return -1;
+    }
+
+    get source() {
+        return this.#source;
+    }
+
+    get timestamp() {
+        return this.#timestamp;
+    }
+
+    get type() {
+        return this.#type;
+    }
+
+    get resolutionIndexes() {
+        return this.#resolutionIndexes;
+    }
+
+    get resolutions() {
+        if (!this.#resolutions && this.#webAppOrtResult) {
+            this.#resolutions = [];
+            this.#resolutionIndexes.forEach((index) => {
+                const webAppResolution = this.#webAppOrtResult.getIssueResolutionByIndex(index) || null;
+                if (webAppResolution) {
+                    this.#resolutions.push(webAppResolution);
+                }
+            });
+        }
+
+        return this.#resolutions;
+    }
+
+    get resolutionReasons() {
+        if (!this.#resolutionReasons && this.#webAppOrtResult) {
+            this.#resolutionReasons = new Set();
+            this.#resolutionIndexes.forEach((index) => {
+                const webAppResolution = this.#webAppOrtResult.getIssueResolutionByIndex(index) || null;
+                if (webAppResolution && webAppResolution.reason) {
+                    this.#resolutionReasons.add(webAppResolution.reason);
+                }
+            });
+        }
+
+        return this.#resolutionReasons;
     }
 }
 
