@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 HERE Europe B.V.
+ * Copyright (C) 2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,28 +18,119 @@
  */
 
 class WebAppPath {
-    #pkg = '';
+    #_id;
 
-    #scope = '';
+    #package;
 
-    constructor(obj) {
+    #packageIndex;
+
+    #project;
+
+    #projectIndex;
+
+    #scope;
+
+    #scopeIndex;
+
+    #path;
+
+    #pathIndexes;
+
+    #webAppOrtResult;
+
+    constructor(obj, webAppOrtResult) {
         if (obj) {
-            if (obj.pkg) {
-                this.#pkg = obj.pkg;
+            if (Number.isInteger(obj._id)) {
+                this.#_id = obj._id;
             }
 
-            if (obj.scope) {
-                this.#scope = obj.scope;
+            if (obj.pkg || obj.pkg >= 0) {
+                this.#packageIndex = obj.pkg;
+            }
+
+            if (obj.project || obj.project >= 0) {
+                this.#projectIndex = obj.project;
+            }
+
+            if (obj.scope || obj.scope >= 0) {
+                this.#scopeIndex = obj.scope;
+            }
+
+            if (obj.path) {
+                this.#pathIndexes = new Set(obj.path);
+            }
+
+            if (webAppOrtResult) {
+                this.#webAppOrtResult = webAppOrtResult;
             }
         }
     }
 
-    get pkg() {
-        return this.#pkg;
+    get _id() {
+        return this.#_id;
+    }
+
+    get package() {
+        if (!this.#package && this.#webAppOrtResult) {
+            const webAppPackage = this.#webAppOrtResult.getPackageByIndex(this.#packageIndex);
+            if (webAppPackage) {
+                this.#package = webAppPackage;
+            }
+        }
+
+        return this.#package;
+    }
+
+    get packageName() {
+        return this.package.id;
+    }
+
+    get project() {
+        if (!this.#project && this.#webAppOrtResult) {
+            const webAppPackage = this.#webAppOrtResult.getPackageByIndex(this.#projectIndex);
+            if (webAppPackage) {
+                this.#project = webAppPackage;
+            }
+        }
+
+        return this.#project;
+    }
+
+    get projectIndex() {
+        return this.#projectIndex;
+    }
+
+    get projectName() {
+        return this.project.id;
     }
 
     get scope() {
+        if (!this.#scope && this.#webAppOrtResult) {
+            const webAppScope = this.#webAppOrtResult.getScopeByIndex(this.#scopeIndex);
+            if (webAppScope) {
+                this.#scope = webAppScope;
+            }
+        }
+
         return this.#scope;
+    }
+
+    get scopeName() {
+        return this.scope.name;
+    }
+
+    get path() {
+        if (!this.#path && this.#pathIndexes && this.#webAppOrtResult) {
+            this.#path = new Set();
+            this.#pathIndexes.forEach((index) => {
+                const webAppPackage = this.#webAppOrtResult.getPackageByIndex(index);
+                if (webAppPackage) {
+                    this.#path.add(webAppPackage);
+                }
+            });
+        }
+
+        return this.#path;
     }
 }
 
