@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 HERE Europe B.V.
+ * Copyright (C) 2020 HERE Europe B.V.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,26 +17,34 @@
  * License-Filename: LICENSE
  */
 
-class ScannerDetails {
+class WebAppScope {
+    #_id;
+
+    #name;
+
     constructor(obj) {
-        this.name = '';
-        this.version = '';
-        this.configuration = '';
+        if (obj) {
+            if (Number.isInteger(obj._id)) {
+                this.#_id = obj._id;
+            }
 
-        if (obj instanceof Object) {
             if (obj.name) {
-                this.name = obj.name;
-            }
-
-            if (obj.version) {
-                this.version = obj.version;
-            }
-
-            if (obj.configuration) {
-                this.configuration = obj.configuration;
+                this.#name = obj.name;
             }
         }
     }
+
+    get _id() {
+        return this.#_id;
+    }
+
+    get id() {
+        return this.#_id;
+    }
+
+    get name() {
+        return this.#name;
+    }
 }
 
-export default ScannerDetails;
+export default WebAppScope;
